# backend/api/swagger.yaml
openapi: "3.0.3"
info:
  title: 算力撮合交易平台 API
  description: |
    ## 概述
    算力现货挂牌撮合交易平台 v1 API。包含账户认证、算力撮合、支付分账、居间金融、运营管理、区块链存证六大模块。

    **Base URL**: `http://localhost:8080/api/v1`

    ## 鉴权
    除公开接口外，所有接口需 `Authorization: Bearer <jwt_token>`

    ## 金额单位
    所有金额单位为**分（fen）**，即 `100 = ¥1.00`

    ## 通用响应格式
    ```json
    { "code": 0, "message": "success", "data": {}, "request_id": "xxx" }
    ```
  version: "1.0.0"
  contact:
    name: Platform Team
servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Response:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: "success" }
        data: { type: object }
        request_id: { type: string }
...
security:
  - BearerAuth: []
paths:
...
  /auth/captcha/verify:
    post:
      tags: [认证 Auth]
      summary: 校验一次性人机验证 token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [captcha_token]
              properties:
                captcha_token: { type: string }
      responses:
        '200':
          description: "请求已处理；token 仅可成功使用一次"
  /auth/sms/code:
    post:
      tags: [认证 Auth]
      summary: 发送登录或注册验证码
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, purpose, captcha_token]
              properties:
                phone: { type: string, example: "13800138000" }
                purpose: { type: string, enum: [register, login], example: "login" }
                captcha_token: { type: string }
      responses:
        '200':
          description: "请求已处理；注册手机号已存在时返回业务码 40900 且不发送验证码；否则验证码有效期默认 300 秒，60 秒后可重新获取"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: success }
                  data:
                    type: object
                    properties:
                      expires_in: { type: integer, example: 300 }
                      resend_after: { type: integer, example: 60 }
                      preview_code:
                        type: string
                        pattern: '^[0-9]{6}$'
                        description: "仅本地 debug preview 模式返回；release 模式禁止启用"
  /auth/sms/login:
    post:
      tags: [认证 Auth]
      summary: 已注册用户短信验证码登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, sms_code]
              properties:
                phone: { type: string, example: "13800138000" }
                sms_code: { type: string, pattern: '^[0-9]{6}$', example: "123456" }
      responses:
        '200':
          description: "登录成功，返回 access_token、refresh_token 与用户信息"
  /auth/register:
    post:
      tags: [认证 Auth]
      summary: 手机号验证码注册并建立会话
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, sms_code, agree_tos]
              properties:
                phone: { type: string, example: "13800138000" }
                sms_code: { type: string, example: "123456" }
                agree_tos: { type: boolean, example: true }
      responses:
        '200':
          description: "注册成功，返回 access_token、refresh_token 与默认 buyer 用户信息"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          access_token: { type: string }
                          refresh_token: { type: string }
                          expires_in: { type: integer, example: 900 }
                          user: { type: object }
  /auth/refresh:
    post:
      tags: [认证 Auth]
      summary: 刷新 Token（refresh token 一次性轮换）
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
  /auth/logout:
    post:
      tags: [认证 Auth]
      summary: 登出
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token: { type: string, description: "传入时与 access token 一并撤销" }
...